% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adap_divide_ssanova.R
\name{adap.divide.ssanova}
\alias{adap.divide.ssanova}
\title{Do adaptive sampling smoothing spline with method of "divide and conquer".}
\usage{
adap.divide.ssanova(x, y, subsetNumber = 1, subsetElementIndex = NULL,
  ratio = NULL, alpha = NULL, nbasis = NULL, nslice = 10,
  sliceMethod = NULL, splineFormula = NULL, pred = TRUE,
  shareBasis = TRUE, uniformSampling = FALSE)
}
\arguments{
\item{x}{A vector (or a matrix or data.frame) of independent varibles.}

\item{y}{A vector of response varibles with noise.}

\item{subsetElementIndex}{A list of indices of each subset's element in whole data.
Default = NULL, randomly generate subsets.}

\item{ratio}{A vector of the ratio of size of each subset. Default is equal.}

\item{alpha}{A number of modifying minimizing GCV in \code{gss::ssanova}.
Default = NULL (-1).}

\item{nbasis}{A number of approximate sample size in adaptive sample.
If it is \code{NULL}, the default number is \code{max(30,
10*n^(2/9))}.}

\item{nslice}{A number of slices numbers. If \code{sliceMethod} is not
\code{NULL}, it will be invaid.}

\item{sliceMethod}{A string of slice method among "Scott","Sturges", "FD"
(Freddman and Diaconis).
The default is \code{NULL} (NULL is not a string).
The number of slices depends on the method.}

\item{splineFormula}{A string of formula of spline function. See more
description in details.}

\item{pred}{A boolean whether return predition value or list of models.}

\item{shareBasis}{A boolean whether each subset shares the same basis.
Default = TRUE: all subsets uses the basis in whole data;
if FALSE: all subsets do adaptive or uniform sampling
themselves.}

\item{uniformSampling}{A boolean whether use uniform sampling basis in each
subset. If FALSE, use adptive sampling.
Require \code{subsetElementIndex != NULL}.}

\item{susetNumber}{A number of subsets. (Denote as m).}
}
\value{
A list:
\itemize{
  \item (all)A number of fitting time of all subset models.
  \item (pred = TRUE)An N*1 vector of preditions of average model
    (the average value of all subset models.)
  \item (pred = TRUE)An N*m matrix of preditions of each subset model.
  \item (pred = FALSE)An list of m fitted models on each subset.
  \item subsetElementIndex: the indices of each subset's elements in whole data
  \item basisIndexInSubset: the indices of basis in each subset
}
}
\description{
Return a list of x and y value generating by 1D FanGijbels function.
}
