% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_sample.R
\name{adap.sample}
\alias{adap.sample}
\title{do adaptive sampling depending on y value}
\usage{
adap.sample(y, nstar = NULL, nslice = 10, sliceMethod = "NULL")
}
\arguments{
\item{y}{A vecotr of response value.}

\item{nstar}{A number of sample size. If it is \code{NULL}, the number would be \eqn{10n^{2/9}}}

\item{nslice}{A number of slices numbers. If \code{sliceMethod} is not \code{NULL}, it isn't effective.}

\item{sliceMethod}{A string of slice method among "NULL", "Scott", "Sturges", "FD"(Freddman and Diaconis)}
}
\value{
Index of selected samples in \code{y}.
}
\description{
do adaptive sampling depending on y value
}
\examples{
adap.sample(y=1:10, sliceMethod="Scott")
}
