% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_ssanova.R
\name{adap.ssanova}
\alias{adap.ssanova}
\title{Fit ssanova with sampling model.}
\usage{
adap.ssanova(x, y, alpha = NULL, nbasis = NULL, nslice = 10,
  sliceMethod = NULL, splineFormula = NULL, returnIndex = FALSE,
  plotFit = FALSE)
}
\arguments{
\item{x}{A vector (or a matrix or data.frame) of independent varibles.
(But now, we just realize vector part)}

\item{y}{A vector of response varibles with noise}

\item{alpha}{A number of modifying minimizing GCV in \code{gss::ssanova}. Default = NULL (1.4).}

\item{nbasis}{A number of approximate sample size in adaptive sample.
If it is \code{NULL}, the default number is \code{max(30, 10*n^(2/9))}.}

\item{nslice}{A number of slices numbers. If \code{sliceMethod} is not \code{NULL}, it will be invaid.}

\item{sliceMethod}{A string of slice method among "Scott","Sturges", "FD"(Freddman and Diaconis).
The default is \code{NULL} (NULL is not a string).
The number of slices depends on the method.}

\item{splineFormula}{A string of formula of spline function. See more description in details.}

\item{returnIndex}{A boolean whether return index of samples.}

\item{plotFit}{A boolean to plot saccter of (x,y), the samples and the smoothing curve.}
}
\value{
A list of fitted model generated by \code{gss::ssanova}.

        If required return index of samples, return a list named \code{fitModel}(list)
        and \code{sample.index}(vector).
}
\description{
Do adaptive sampling and then fit ssanova model with Gusssian noise.
Return the fitmodel of function \code{gss::ssanova}
}
\details{
If x is a matrix, generate a data.frame of x & y;
  generate string of formular for high demesion like: \code{"y ~ x1 * x2 * ... * xn"}.
  Do not plot if x is a matrix.

  When use function: \code{predict()}, follow the naming scheme as "x1, x2, ... xn"
  in new data.frame

  In parameter \code{splineFormula}, default is \code{NULL}.
  \itemize{
    \item If x is 1D, default is \code{"y~x"};
    \item If x is mD, default is \code{"y ~ x1*x2*... *xm"}.
    \item \code{"1 +x1 + x2"} stands for:      c + f1(x1) + f2(x2);
    \item \code{"1 + x1 + x1:x2"} stands for:  c + f1(x1) + f12(x1,x2);
    \item \code{"1 + x1*x2"} stands for:       c + f1(x1) + f2(x2) + f12(x1,x2);
    \item etc.
  }
}
\seealso{
\code{\link{adap.sample}} for details of adaptive sampling.
}
