% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocateToSlice.R
\name{allocateToSlice}
\alias{allocateToSlice}
\title{allocate elements into different slices}
\usage{
allocateToSlice(nelement, nslice)
}
\arguments{
\item{nelement}{The number of elements to be allocated.}

\item{nslice}{The number of slices.}
}
