% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJ_EX_rand.R
\name{DJ.EX.rand}
\alias{DJ.EX.rand}
\title{Generate the all functions in Donoho and Johnstone(1994)}
\usage{
DJ.EX.rand(x, signal = 7, rsnr = 7, noisy = FALSE, plotfn = FALSE)
}
\arguments{
\item{x}{A vector of independent varible.}

\item{signal}{A number about amplitude in response value in "DJ" method.}

\item{rsnr}{A number of relative signal to noise ration.
The standard deviation of noise is \eqn{signal / rsnr}.
If \code{noisy == FALSE}, \code{rsnr} does not work.}

\item{noisy}{A boolean whether add noise to model value.}

\item{plotfn}{A boolean whether plot 4 plots of data.}
}
\value{
A list of 4 vectors of response varibles named "blocks", "bumps", "heavi", "dopper".
}
\description{
generate 4 kinds of responsive varible based on \code{x}:
  "blocks", "bumps", "heavi", "dopper".
  We use the return values for function:\code{\link{gen.simu.data}}.
}
\examples{
# genertae data based on x. Amplitude is 7, do not add noise.
x <- runif(100, 0, 1)
data.list <- DJ.EX.rand(x, signal = 7, nosiy = FALSE)

}
