% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_boundary.R
\name{find.boundary}
\alias{find.boundary}
\title{Find boundary of data}
\usage{
find.boundary(x)
}
\arguments{
\item{x}{A vector, matrix or data.frame of independent varibles.}
}
\value{
A vector of index of boundary cases.
}
\description{
Find boundary of independent varibles, return index of these boundary.

  E.g., when x is \code{c(0, 0.5, 0.2, 0.1)}, the function would return \code{c(1, 2)}.
}
\examples{
# find indices of bounary:
## 1) of a vector:
find.boundary(x = c(0, 0.5, 0.2, 0.1))

## 2) of a matrix:
find.boundary(x = matrix (1:10, nrow=5, ncol=2))


}
