% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_plot.R
\name{fit.plot}
\alias{fit.plot}
\title{Plot fitted model}
\usage{
fit.plot(x, y, v = NULL, sampleIndex = NULL, fitModel = NULL,
  figTitle = "Figure")
}
\arguments{
\item{x}{A vector (or a matrix) of independent varibles.}

\item{y}{A vector of response varibles with noise.}

\item{v}{A vector of response varibles of true value (without noise).}

\item{sampleIndex}{A vector of index of selected sample.}

\item{fitModel}{A list of fitted model of \code{gss::ssanova}.
(result of other fitting function like \code{lm} is also acceptable).}

\item{figTitle}{A string of figure' s title.}
}
\description{
\itemize{
  \item Plot in one graph of original data without noise(if there is),
    data with noise, fitted data.
  \item When fitted model is given, generate 10 times unif points to plot
    a smoothing curve.
  \item \strong{Attetion}: x and y shoule be vector(Just plot a 2D figure).
    The varible name in model must exact be "x" and "y", or it would be
    error in \code{predict}.
    (generate 10 times points by \code{predict} to plot a smoothing curve)
}
}
\examples{
# do linear regression of y~x, and plot the scatter with fitted line on one figure.
# all x and y are selected into regression, so sampleIndex is a vector of all indices.
x <- 1:10
y <- 3 * x + rnorm(10,0,1)
lm.model <- lm(y ~ x)
fitplot(x = x, y= y, sampleIndex = 1:10, fitModel = lm.model,
        figTitle = "an example of lm")

}
