% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_simu_data.R
\name{gen.1dfan.simu.data}
\alias{gen.1dfan.simu.data}
\title{Generate 1D Fan-Gijbels simulation data}
\usage{
gen.1dfan.simu.data(nobs = 1024, nbasis = NULL, nslices = 11,
  sliceMethod = NULL, SNR = 7, iter = 1, test = FALSE)
}
\arguments{
\item{nobs}{A number of observations in simulation.}

\item{nbasis}{A number of approximate sample size. If it is \code{NULL}, the number would be \eqn{10*n^{2/9}}}

\item{sliceMethod}{A string of slice method among "NULL", "Scott",
"Sturges", "FD"(Freddman and Diaconis).}

\item{SNR}{A number of "signal to noise" ratio: the ratio of standard deviation of signal to noise.}

\item{iter}{A number helping record random seeds when testing}

\item{test}{A boolean. If \code{TRUE}, set random seeds.}

\item{nslice}{A number of slices numbers.
If \code{sliceMethod} is not \code{NULL}, it will be invaid.}
}
\description{
Return a list of x and y value generating by 1D FanGijbels function.
}
