% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_formula_and_df.R
\name{gen.formula.and.df}
\alias{gen.formula.and.df}
\title{Generate data.frame and formula for high dimension spline}
\usage{
gen.formula.and.df(x, y)
}
\arguments{
\item{x}{A vector or matrix or data.frame of independent varibles.}

\item{y}{A vector of response varible.}
}
\value{
A list of 1) a data.frame including x and y; 2) a string of
  spline formula; 3) a number of colume rank of x.

  \code{list(xy.df = xy.df, spline.formula = spline.formula, x.dim = x.dim)}
}
\description{
In case x is a matrix or data.frame, generate a data.frame of x & y;
  The columns of new data.frame are named as "x1", "x2", ... "xn", "y".
  generate string of formular for high demesion like: \code{"y ~ x1 * x2
  * ... * xn"}.
}
\examples{
 x <- matrix(1:10, nrow=5, ncol=2)
 y <- 1:10
 result.list <- gen.formula.and.df(x, y)

}
