% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_simu_data.R
\name{gen.simu.data}
\alias{gen.simu.data}
\title{Generate 1D Fan-Gijbels simulation data}
\usage{
gen.simu.data(nobs = 1024, genMethod = "1dfan", SNR = 7, signal = 7,
  seed = NULL)
}
\arguments{
\item{nobs}{A number of observations in simulation.}

\item{genMethod}{A string of data generating method among "1dfan",
"DJ.blocks", "DJ.bumps", "DJ.heavi", "DJ.doppler".}

\item{SNR}{A number of "signal to noise" ratio: the ratio of standard
deviation of signal to noise.}

\item{signal}{A number about amplitude in response value of "DJ" method.}

\item{seed}{A number of base of random seed.}
}
\value{
A list contains generated \code{x, y, v}.
}
\description{
Return a list of x and y value generating by 1D FanGijbels
  function.
}
\details{
In \code{genMethod}: "DJ" stands for functions in Donoho and
  Johnstone(1994).

  The return value of function is a list of \code{x, y, z}.
\itemize{
  \item x: independent varible.
  \item y: response varible with noise.
  \item v: the true value of response varible without noise.
}
}
\examples{
# generate data of "DJ.blocks".
## numbers of observation: 1024, signal to noise ratio is 7.
data.xyv <- gen.simu.data(nobs = 1024, genMethod = "DJ.blocks", SNR = 7)

}
\seealso{
4 kinds of functions of in Donoho and Johnstone(1994) in function:
  \code{\link{DJ.EX.rand}}.
  Add Guassian noise to response varible in function:
  \code{\link{add.noise}}.
}
