% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_subset.R
\name{generate.subset}
\alias{generate.subset}
\title{Generate sampling basis in each subset}
\usage{
generate.subset(N, sample.index, subset.number = 1, ratio = NULL)
}
\arguments{
\item{N}{The number of whole smaples.}

\item{sample.index}{The indices of samples.}

\item{subset.number}{The number of subsets.}

\item{ratio}{A vector of the ratio of size of each subset. Default is equal.}
}
\value{
A list contains two vectors. The indices of samples and basis in each subset.
}
\description{
The subsets is divied randomly and evenly. Return a list of indices of samples and indices basis in each subset.
}
\examples{
generate.subset(y = 1:10, sample.index = c(1,3,5), subset.number = 2)

}
