% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adap_divide_ssanova.R
\name{adap.divide.ssanova}
\alias{adap.divide.ssanova}
\title{Do adaptive sampling smoothing spline with method of "divide and conquer".}
\usage{
adap.divide.ssanova(x, y, subsetNumber = 1, ratio = NULL, alpha = NULL,
  nbasis = NULL, nslice = 10, sliceMethod = NULL, splineFormula = NULL,
  pred = TRUE, predAverage = TRUE, calculateTime = FALSE)
}
\arguments{
\item{x}{A vector (or a matrix or data.frame) of independent varibles.
(But now, we just realize vector part)}

\item{y}{A vector of response varibles with noise}

\item{ratio}{A vector of the ratio of size of each subset. Default is equal.}

\item{alpha}{A number of modifying minimizing GCV in \code{gss::ssanova}. Default = NULL (-1).}

\item{nbasis}{A number of approximate sample size in adaptive sample.
If it is \code{NULL}, the default number is \code{max(30, 10*n^(2/9))}.}

\item{nslice}{A number of slices numbers. If \code{sliceMethod} is not \code{NULL}, it will be invaid.}

\item{sliceMethod}{A string of slice method among "Scott","Sturges", "FD"(Freddman and Diaconis).
The default is \code{NULL} (NULL is not a string).
The number of slices depends on the method.}

\item{splineFormula}{A string of formula of spline function. See more description in details.}

\item{pred}{A boolean whether return predition value or list of models.}

\item{predAverage}{A boolean whether predition an average of all subsets models.
If \code{FALSE}, do not avearage, return the each predition.}

\item{calculateTime}{A boolean whether calculate time of fitting model.}

\item{susetNumber}{A number of subsets. (Denote as m).}
}
\value{
A list contains one or two items as below (depends on parameters):
\itemize{
  \item A number of fitting time of all subset models.
  \item An N*1 vector of preditions of average model (the average value of all subset models.)
  \item An N*m matrix of preditions of each subset model.
  \item An list of m fitted models on each subset.
}
}
\description{
Return a list of x and y value generating by 1D FanGijbels function.
}
