% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_sample.R
\name{adap.sample}
\alias{adap.sample}
\title{Adaptive sampling}
\usage{
adap.sample(y, nbasis = NULL, nslice = 10, sliceMethod = NULL)
}
\arguments{
\item{y}{A vecotr of response value.}

\item{nbasis}{A number of approximate sample size. If it is \code{NULL},
the number would be \code{max(30, 10*(n)^(2/9))}.}

\item{nslice}{A number of slices numbers. If \code{sliceMethod} is not
\code{NULL}, it will be invaid.}

\item{sliceMethod}{A string of slice method among "Scott", "Sturges",
"FD"(Freddman and Diaconis).
The default is \code{NULL} (NULL is not a string).
The number of slices depends on the method in function:
\code{\link[stats]{hist}}.}
}
\value{
A vector of index of selected samples in \code{y}.
}
\description{
Do adaptive sampling depending on y value
}
\examples{
# do adaptive sampling based on the value of y.
## number of samples is default as max(30, 10*n^(2/9))
## slices are divided by rule of "Scoot" of hist().
adap.sample(y=1:100, sliceMethod="Scott")

}
\seealso{
The methods of divided slices in \code{\link[stats]{hist}}.

Get the boundary of x data. \code{\link{find.boundary}}.
}
