% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_subset.R
\name{generate.subset}
\alias{generate.subset}
\title{Generate sampling basis in each subset}
\usage{
generate.subset(y, sample.index, subset.number = 1)
}
\arguments{
\item{y}{The response value in spline.}

\item{sample.index}{The indices of samples.}

\item{subset.number}{The number of subsets.}
}
\value{
A list contains two vectors. The indices of samples and basis in each subset.
}
\description{
The subsets is divied randomly and evenly. Return a list of indices of samples and indices basis in each subset.
}
\examples{
generate.subset(y = 1:10, sample.index = c(1,3,5), subset.number = 2)

}
